Feature/performance refactor (#77)
* refactor: BaseTimeEntity 도입

* post패키지 리팩토링 중간 커밋

* 중형 자동테스트 수정 원인: 레디스 연결실패 이슈

+++ src/main/java/com/kustaurant/kustaurant/admin/report/ReportController.java
@@ -29,7 +29,7 @@ public class ReportController {
     })
     @PostMapping("/api/v1/auth/restaurants/{restaurantId}/comments/{commentId}/report")
     public ResponseEntity<Void> reportComment(
-            @PathVariable Integer restaurantId,
+            @PathVariable Long restaurantId,
             @PathVariable Integer commentId,
             @Parameter(hidden = true) @AuthUser AuthUserInfo user
     ) {
