+++ src/main/java/com/kustaurant/kustaurant/evaluation/comment/service/EvalCommCommandServiceImpl.java
@@ -7,7 +7,7 @@ import com.kustaurant.kustaurant.evaluation.comment.controller.request.EvalComme
 import com.kustaurant.kustaurant.evaluation.comment.domain.EvalComment;
 import com.kustaurant.kustaurant.evaluation.comment.infrastructure.repo.jpa.EvalCommUserReactionRepository;
 import com.kustaurant.kustaurant.evaluation.comment.service.port.EvalCommentRepository;
-import com.kustaurant.kustaurant.global.exception.exception.business.DataNotFoundException;
+import com.kustaurant.kustaurant.global.exception.exception.DataNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,12 +19,12 @@ public class EvalCommCommandServiceImpl implements EvalCommCommandService {
     private final EvalCommentRepository evalCommentRepository;
     private final EvalCommUserReactionRepository evalCommUserReactionRepository;
 
-    public EvalComment create(Long evaluationId, Integer restaurantId, Long userId, EvalCommentRequest req) {
+    public EvalComment create(Long evaluationId, Long restaurantId, Long userId, EvalCommentRequest req) {
         EvalComment evalComment = EvalComment.create(userId, restaurantId, evaluationId, req);
         return evalCommentRepository.save(evalComment);
     }
 
-    public void delete(Long evalCommentId, Integer restaurantId, Long userId) {
+    public void delete(Long evalCommentId, Long restaurantId, Long userId) {
         EvalComment evalComment = evalCommentRepository.findByIdAndRestaurantId(evalCommentId, restaurantId)
                 .orElseThrow(() -> new DataNotFoundException(COMMENT_NOT_FOUND));
         evalComment.softDelete(userId);
