+++ src/main/java/com/kustaurant/kustaurant/evaluation/evaluation/infrastructure/EvaluationQueryRepositoryImpl.java
@@ -7,8 +7,7 @@ import com.kustaurant.kustaurant.evaluation.evaluation.infrastructure.entity.Eva
 import com.kustaurant.kustaurant.evaluation.evaluation.infrastructure.jpa.EvaluationQueryJpaRepository;
 import com.kustaurant.kustaurant.evaluation.evaluation.service.port.EvaluationQueryRepository;
 
-import com.kustaurant.kustaurant.global.exception.ErrorCode;
-import com.kustaurant.kustaurant.global.exception.exception.business.DataNotFoundException;
+import com.kustaurant.kustaurant.global.exception.exception.DataNotFoundException;
 import java.util.List;
 import java.util.Optional;
 
@@ -29,7 +28,7 @@ public class EvaluationQueryRepositoryImpl implements EvaluationQueryRepository 
     }
 
     @Override
-    public boolean existsByUserAndRestaurant(Long userId, Integer restaurantId) {
+    public boolean existsByUserAndRestaurant(Long userId, Long restaurantId) {
         if (userId == null || restaurantId == null) {
             return false;
         }
@@ -37,7 +36,7 @@ public class EvaluationQueryRepositoryImpl implements EvaluationQueryRepository 
     }
 
     @Override
-    public boolean existsByRestaurantAndEvaluation(Integer restaurantId, Long evaluationId) {
+    public boolean existsByRestaurantAndEvaluation(Long restaurantId, Long evaluationId) {
         if (restaurantId == null || evaluationId == null) {
             return false;
         }
@@ -45,7 +44,7 @@ public class EvaluationQueryRepositoryImpl implements EvaluationQueryRepository 
     }
 
     @Override
-    public Optional<Evaluation> findActiveByUserAndRestaurant(Long userId, Integer restaurantId) {
+    public Optional<Evaluation> findActiveByUserAndRestaurant(Long userId, Long restaurantId) {
         return jpaRepository.findByUserIdAndRestaurantIdAndStatus(userId, restaurantId, "ACTIVE")
                 .map(EvaluationEntity::toModel);
     }
@@ -56,7 +55,7 @@ public class EvaluationQueryRepositoryImpl implements EvaluationQueryRepository 
     }
 
     @Override
-    public List<Evaluation> findByRestaurantIdOrderByCreatedAtDesc(Integer restaurantId) {
+    public List<Evaluation> findByRestaurantIdOrderByCreatedAtDesc(Long restaurantId) {
         return jpaRepository.findByRestaurantIdAndStatusOrderByCreatedAtDesc(restaurantId, "ACTIVE")
                 .stream()
                 .map(EvaluationEntity::toModel)
@@ -64,7 +63,7 @@ public class EvaluationQueryRepositoryImpl implements EvaluationQueryRepository 
     }
 
     @Override
-    public List<Evaluation> findByRestaurantIdOrderByLikeCountDesc(Integer restaurantId) {
+    public List<Evaluation> findByRestaurantIdOrderByLikeCountDesc(Long restaurantId) {
         return jpaRepository.findByRestaurantIdAndStatusOrderByLikeCountDesc(restaurantId, "ACTIVE")
                 .stream()
                 .map(EvaluationEntity::toModel)
