+++ src/main/java/com/kustaurant/kustaurant/evaluation/evaluation/infrastructure/jpa/EvaluationQueryJpaRepository.java
@@ -10,16 +10,16 @@ public interface EvaluationQueryJpaRepository extends Repository<EvaluationEntit
 
     Optional<EvaluationEntity> findByIdAndStatus(Long id, String status);
 
-    boolean existsByUserIdAndRestaurantId(Long userId, Integer restaurantId);
+    boolean existsByUserIdAndRestaurantId(Long userId, Long restaurantId);
 
-    boolean existsByRestaurantIdAndId(Integer restaurantId, Long evaluationId);
+    boolean existsByRestaurantIdAndId(Long restaurantId, Long evaluationId);
 
-    Optional<EvaluationEntity> findByUserIdAndRestaurantIdAndStatus(Long userId, Integer restaurantId, String status);
+    Optional<EvaluationEntity> findByUserIdAndRestaurantIdAndStatus(Long userId, Long restaurantId, String status);
 
     Integer countByStatus(String status);
 
-    List<EvaluationEntity> findByRestaurantIdAndStatusOrderByCreatedAtDesc(Integer restaurantId, String status);
+    List<EvaluationEntity> findByRestaurantIdAndStatusOrderByCreatedAtDesc(Long restaurantId, String status);
 
-    List<EvaluationEntity> findByRestaurantIdAndStatusOrderByLikeCountDesc(Integer restaurantId, String status);
+    List<EvaluationEntity> findByRestaurantIdAndStatusOrderByLikeCountDesc(Long restaurantId, String status);
 
 }
