+++ src/main/java/com/kustaurant/kustaurant/rating/domain/model/RatingScore.java
@@ -1,7 +1,7 @@
 package com.kustaurant.kustaurant.rating.domain.model;
 
 public record RatingScore(
-        int restaurantId,
+        long restaurantId,
         double score
 ) {
