+++ src/main/java/com/kustaurant/kustaurant/restaurant/favorite/model/RestaurantFavorite.java
@@ -11,12 +11,12 @@ public class RestaurantFavorite {
 
     private Integer id;
     private Long userId;
-    private Integer restaurantId;
+    private Long restaurantId;
     private String status;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
-    public static RestaurantFavorite create(Long userId, Integer restaurantId) {
+    public static RestaurantFavorite create(Long userId, Long restaurantId) {
         return RestaurantFavorite.builder()
                 .userId(userId)
                 .restaurantId(restaurantId)
