+++ src/main/java/com/kustaurant/kustaurant/restaurant/query/common/infrastructure/query/RestaurantCoreInfoQuery.java
@@ -33,8 +33,8 @@ public class RestaurantCoreInfoQuery {
      * 식당 id들에 대해 식당 요약 정보(RestaurantCoreInfoDto)를 반환함.
      * 기본적인 식당 정보 + (userId가 존재하면) 평가 여부 + 즐찾 여부 등
      */
-    public List<RestaurantCoreInfoDto> getRestaurantTiers(List<Integer> restaurantIds, Long userId) {
-        Map<Integer, RestaurantCoreInfoDto> map = queryFactory
+    public List<RestaurantCoreInfoDto> getRestaurantTiers(List<Long> restaurantIds, Long userId) {
+        Map<Long, RestaurantCoreInfoDto> map = queryFactory
                 .from(restaurantEntity)
                 .leftJoin(ratingEntity).on(ratingEntity.restaurantId.eq(restaurantEntity.restaurantId))
                 .leftJoin(restaurantSituationRelationEntity)
@@ -74,12 +74,12 @@ public class RestaurantCoreInfoQuery {
                 .toList();
     }
 
-    private BooleanExpression evaluationRestaurantIdEq(NumberPath<Integer> restaurantId, Long userId) {
+    private BooleanExpression evaluationRestaurantIdEq(NumberPath<Long> restaurantId, Long userId) {
         return isNull(userId) ? Expressions.FALSE
                 : evaluationEntity.restaurantId.eq(restaurantId).and(evaluationEntity.userId.eq(userId));
     }
 
-    private BooleanExpression favoriteRestaurantIdEq(NumberPath<Integer> restaurantId, Long userId) {
+    private BooleanExpression favoriteRestaurantIdEq(NumberPath<Long> restaurantId, Long userId) {
         return isNull(userId) ? Expressions.FALSE
                 : restaurantFavoriteEntity.restaurantId.eq(restaurantId).and(restaurantFavoriteEntity.userId.eq(userId));
     }
