+++ src/main/java/com/kustaurant/kustaurant/restaurant/query/draw/RestaurantDrawService.java
@@ -2,9 +2,8 @@ package com.kustaurant.kustaurant.restaurant.query.draw;
 
 import static com.kustaurant.kustaurant.global.exception.ErrorCode.*;
 
-import com.kustaurant.kustaurant.global.exception.exception.business.DataNotFoundException;
+import com.kustaurant.kustaurant.global.exception.exception.DataNotFoundException;
 import com.kustaurant.kustaurant.restaurant.query.common.dto.RestaurantCoreInfoDto;
-import com.kustaurant.kustaurant.restaurant.query.chart.service.port.RestaurantChartRepository;
 import com.kustaurant.kustaurant.restaurant.restaurant.domain.Restaurant;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,7 +20,7 @@ public class RestaurantDrawService {
 
     Random rand = new Random();
 
-    public Restaurant getById(Integer restaurantId) {
+    public Restaurant getById(Long restaurantId) {
         return restaurantDrawRepository.getById(restaurantId);
     }
